requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
aiogram==3.4.1
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
python-multipart==0.0.6
jinja2==3.1.3
httpx==0.25.2
python-dotenv==1.0.0
